feat: Add endpoint for adding authors
This required changes to the underlying AuthorsModel. It now
reads and writes directly from the Authors.json file, and does not
store a List of Authors in memory. This is to ensure that the data
is persisted between restarts of the application, and also to ensure
that if an AuthorsModel instance is dropped from memory, it will
not lose any data.